+++ mod.ts
@@ -7,6 +7,7 @@ import setTag from "./plugins/set.ts";
 import jsTag from "./plugins/js.ts";
 import layoutTag from "./plugins/layout.ts";
 import functionTag from "./plugins/function.ts";
+import importTag from "./plugins/import.ts";
 import escape from "./plugins/escape.ts";
 
 export interface Options {
@@ -31,6 +32,7 @@ export default function (options: Options = {}) {
   env.use(includeTag());
   env.use(setTag());
   env.use(layoutTag());
+  env.use(importTag());
   env.use(functionTag());
   env.use(escape());