+++ plugins/import.ts
@@ -1,0 +1,32 @@
+import type { Environment } from "../src/environment.ts";
+
+export default function () {
+  return (env: Environment) => {
+    env.tags.push(importTag);
+  };
+}
+
+function importTag(
+  _env: Environment,
+  code: string,
+): string | undefined {
+  if (!code.startsWith("import ")) {
+    return;
+  }
+
+  const match = code?.match(
+    /^import\s+(\{[\s|\S]*\})\s+from\s+(.+)$/,
+  );
+
+  if (!match) {
+    throw new Error(`Invalid import: ${code}`);
+  }
+
+  const [_, vars, file] = match;
+
+  return `
+    __imports = {};
+    await __env.run(${file}, {...__data}, __file, __imports);
+    let ${vars} = __imports;
+  `;
+}