+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [0.7.0] - Unreleased
 ### Added
 - `function` tag [#9].
+- `import` and `export` tag [#9].
 - `runStringSync` function to run Vento in sync mode.
 
 ## [0.6.0] - 2023-07-03